# Use Bytes instead of &[u8] to reduce memory copying