--- a/sdk/src/binary/consumer_offsets.rs
+++ b/sdk/src/binary/consumer_offsets.rs
@@ -12,7 +12,7 @@ use crate::models::consumer_offset_info::ConsumerOffsetInfo;
 impl<B: BinaryClient> ConsumerOffsetClient for B {
     async fn store_consumer_offset(&self, command: &StoreConsumerOffset) -> Result<(), IggyError> {
         fail_if_not_authenticated(self).await?;
-        self.send_with_response(STORE_CONSUMER_OFFSET_CODE, &command.as_bytes())
+        self.send_with_response(STORE_CONSUMER_OFFSET_CODE, command.as_bytes())
             .await?;
         Ok(())
     }
@@ -23,8 +23,8 @@ impl<B: BinaryClient> ConsumerOffsetClient for B {
     ) -> Result<ConsumerOffsetInfo, IggyError> {
         fail_if_not_authenticated(self).await?;
         let response = self
-            .send_with_response(GET_CONSUMER_OFFSET_CODE, &command.as_bytes())
+            .send_with_response(GET_CONSUMER_OFFSET_CODE, command.as_bytes())
             .await?;
-        mapper::map_consumer_offset(&response)
+        mapper::map_consumer_offset(response)
     }
 }