--- a/sdk/src/binary/streams.rs
+++ b/sdk/src/binary/streams.rs
@@ -20,43 +20,43 @@ impl<B: BinaryClient> StreamClient for B {
     async fn get_stream(&self, command: &GetStream) -> Result<StreamDetails, IggyError> {
         fail_if_not_authenticated(self).await?;
         let response = self
-            .send_with_response(GET_STREAM_CODE, &command.as_bytes())
+            .send_with_response(GET_STREAM_CODE, command.as_bytes())
             .await?;
-        mapper::map_stream(&response)
+        mapper::map_stream(response)
     }
 
     async fn get_streams(&self, command: &GetStreams) -> Result<Vec<Stream>, IggyError> {
         fail_if_not_authenticated(self).await?;
         let response = self
-            .send_with_response(GET_STREAMS_CODE, &command.as_bytes())
+            .send_with_response(GET_STREAMS_CODE, command.as_bytes())
             .await?;
-        mapper::map_streams(&response)
+        mapper::map_streams(response)
     }
 
     async fn create_stream(&self, command: &CreateStream) -> Result<(), IggyError> {
         fail_if_not_authenticated(self).await?;
-        self.send_with_response(CREATE_STREAM_CODE, &command.as_bytes())
+        self.send_with_response(CREATE_STREAM_CODE, command.as_bytes())
             .await?;
         Ok(())
     }
 
     async fn delete_stream(&self, command: &DeleteStream) -> Result<(), IggyError> {
         fail_if_not_authenticated(self).await?;
-        self.send_with_response(DELETE_STREAM_CODE, &command.as_bytes())
+        self.send_with_response(DELETE_STREAM_CODE, command.as_bytes())
             .await?;
         Ok(())
     }
 
     async fn update_stream(&self, command: &UpdateStream) -> Result<(), IggyError> {
         fail_if_not_authenticated(self).await?;
-        self.send_with_response(UPDATE_STREAM_CODE, &command.as_bytes())
+        self.send_with_response(UPDATE_STREAM_CODE, command.as_bytes())
             .await?;
         Ok(())
     }
 
     async fn purge_stream(&self, command: &PurgeStream) -> Result<(), IggyError> {
         fail_if_not_authenticated(self).await?;
-        self.send_with_response(PURGE_STREAM_CODE, &command.as_bytes())
+        self.send_with_response(PURGE_STREAM_CODE, command.as_bytes())
             .await?;
         Ok(())
     }