--- a/sdk/src/bytes_serializable.rs
+++ b/sdk/src/bytes_serializable.rs
@@ -1,12 +1,14 @@
+use bytes::Bytes;
+
 use crate::error::IggyError;
 
 /// The trait represents the logic responsible for serializing and deserializing the struct to and from bytes.
 pub trait BytesSerializable {
     /// Serializes the struct to bytes.
-    fn as_bytes(&self) -> Vec<u8>;
+    fn as_bytes(&self) -> Bytes;
 
     /// Deserializes the struct from bytes.
-    fn from_bytes(bytes: &[u8]) -> Result<Self, IggyError>
+    fn from_bytes(bytes: Bytes) -> Result<Self, IggyError>
     where
         Self: Sized;
 }