… sdk/src/binary/topics.rs
… sdk/src/binary/users.rs
… sdk/src/consumer_groups/create_consumer_group.rs
… sdk/src/consumer_groups/delete_consumer_group.rs
… sdk/src/consumer_groups/get_consumer_group.rs
… sdk/src/consumer_groups/get_consumer_groups.rs
… sdk/src/consumer_groups/join_consumer_group.rs
… sdk/src/consumer_groups/leave_consumer_group.rs
… sdk/src/consumer_offsets/get_consumer_offset.rs
… sdk/src/consumer_offsets/store_consumer_offset.rs
… sdk/src/identifier.rs
… sdk/src/messages/poll_messages.rs
… sdk/src/messages/send_messages.rs
… sdk/src/models/header.rs
… sdk/src/models/messages.rs
… sdk/src/models/permissions.rs
… sdk/src/partitions/create_partitions.rs
… sdk/src/partitions/delete_partitions.rs
… sdk/src/personal_access_tokens/create_personal_access_token.rs
… sdk/src/personal_access_tokens/delete_personal_access_token.rs
… sdk/src/personal_access_tokens/get_personal_access_tokens.rs
… sdk/src/personal_access_tokens/login_with_personal_access_token.rs
… sdk/src/quic/client.rs
… sdk/src/streams/create_stream.rs
… sdk/src/streams/delete_stream.rs
… sdk/src/streams/get_stream.rs
… sdk/src/streams/get_streams.rs
… sdk/src/streams/purge_stream.rs
… sdk/src/streams/update_stream.rs
… sdk/src/system/get_client.rs