When using &[u8], it's hard to avoid memory copying due to lifetime issues. Replacing &[u8] with Bytes can effectively reduce memory copying.