@hubcio 2024-01-29
for me server panics:
```
thread 'tokio-runtime-worker' panicked at sdk/src/command.rs:261:47:
range end index 4 out of range for slice of length 0
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

when running ` target/release/iggy-bench -v send tcp`