@YHM404 2024-01-30
> please run bench and attach before/after to the PR

This PR is still in progress. 
There are strange performance issues.