@YHM404 2024-01-30
**Before**
Send TCP
```
2024-01-30T17:40:44.088953Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 1120.89 MB/s, 1120894 messages/s, average latency: 8.92 ms, average  throughput: 112.09 MB/s, total duration: 8.92 s
```

Poll TCP
```
2024-01-30T17:41:48.839996Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 2108.67 MB/s, 2017868 messages/s, average latency: 4.85 ms, average  throughput: 210.87 MB/s, total duration: 4.96 s
```

Send && poll TCP
```
2024-01-30T17:40:10.334675Z  INFO iggy_bench::benchmark_runner: Producer results: total throughput: 1822.24 MB/s, 1822238 messages/s, average latency: 5.49 ms, average throughput: 91.11 MB/s, total duration: 5.49 s, Consumer results: total throughput: 1882.17 MB/s, 1801123 messages/s, average latency: 5.36 ms, average  throughput: 94.11 MB/s, total duration: 5.55 s
2024-01-30T17:40:10.334678Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 1852.38 MB/s, 1811619 messages/s, average latency: 10.85 ms, average  throughput: 92.62 MB/s, total duration: 11.04 s
```

**After**
Send TCP
```
2024-01-30T17:32:46.093686Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 1078.68 MB/s, 1078682 messages/s, average latency: 9.27 ms, average  throughput: 107.87 MB/s, total duration: 9.27 s
```

poll TCP
```
2024-01-30T17:33:20.772766Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 2026.42 MB/s, 1939158 messages/s, average latency: 5.05 ms, average  throughput: 202.64 MB/s, total duration: 5.16 s
```

Send && poll TCP
```
2024-01-30T17:35:47.037283Z  INFO iggy_bench::benchmark_runner: Producer results: total throughput: 2208.14 MB/s, 2208138 messages/s, average latency: 4.53 ms, average throughput: 110.41 MB/s, total duration: 4.53 s, Consumer results: total throughput: 2232.02 MB/s, 2135903 messages/s, average latency: 4.57 ms, average  throughput: 111.60 MB/s, total duration: 4.68 s
2024-01-30T17:35:47.037285Z  INFO iggy_bench::benchmark_runner: Results: total throughput: 2220.28 MB/s, 2171420 messages/s, average latency: 9.09 ms, average  throughput: 111.01 MB/s, total duration: 9.21 s
```